+++ src/utility.h
@@ -1,17 +1,19 @@
 #ifndef UTILITY_H
 #define UTILITY_H
 
+#include <stdint.h>
+
 namespace PixelMaestro {
 	class Utility {
 		public:
-			static unsigned int abs_int(short val);
-			static int rand(int max = 0);
+			static uint32_t abs_int(int32_t val);
+			static int32_t rand(int32_t max = 0);
 			static float sqrt(float val);
-			static unsigned int square(short val);
+			static uint32_t square(int16_t val);
 
 		private:
 			/// Initial RNG seed.
-			static int seed_;
+			static int32_t seed_;
 	};
 }
